【重要更新】：将jdk更换为17 新增方法： +管理员新增获取rsa公私钥方法 +商品新增id数组查询多个商品方法 +用户新增忘记密码方法 修改方法： *图片上传类去除上传图片名需要在20字数限制内 *JWT里去掉不必要的字段 *订单提交方法顺便对销量进行加一 新增注解： @EncryptResponse（此注解用于方法上面进行标记是否加密有就加密数据） 新增类： ^EncryptionFilter，此类用于解密请求的加密数据 ^EncryptionResponseBodyAdvice，因为请求体只能被读取一次，所以此类用于将解密的数据返回请求体中 ^RSAKeyPairGenerator，此类用于加密数据 ^GlobalSettings，此类用于对网站进行动态的全局管理，如是否显示侧边栏等

+++ src/main/java/com/buy_customers/service/impl/ProductServiceImpl.java
@@ -61,7 +61,7 @@ public class ProductServiceImpl extends ServiceImpl<ProductDao, Product> impleme
         return iPage;
     }
     @Override
-    public boolean updateProduct(Integer productId, String productName, Integer price, String category) {
+    public boolean updateProduct(Integer productId, String productName, Double price, String category) {
         Product product = new Product();
         product.setProductId(productId);
         if (productName!=null&& !productName.isEmpty()){
